requirements.txt
aiohttp==3.9.2
asyncio==3.4.3
discord.py==1.7.3
twilio==8.10.0
vonage==3.8.0
plivo==4.30.0
clicksend-client==5.0.37
messagebird==1.5.0
sinch-sdk==0.10.0
